Refactor: create YamlParser and Localizer classes.
This separates YAML parsin and Localizer into YamlParser and Localizer,
each compiled into their own objects. This also adds YAML parsing timing
information.
